Initial Rough
This is python 3.6 program that takes two arguments (a PDF and a declination). It will then draw lines over the PDF at the desired declination. This is to help with compasses that do not have the ability to adjust declination.

Initial maps PDFs are gathered from CalTopo for my testing purposes.